QSoundEffect: Check if network access manager is already released
In case if network access manager is already released need to prevent
using it in dtor.

Task-number: QTBUG-66754

+++ src/multimedia/audio/qsamplecache_p.cpp
@@ -132,7 +132,8 @@ QSampleCache::~QSampleCache()
     for (QSample* sample : copyStaleSamples)
         delete sample;
 
-    m_networkAccessManager->deleteLater();
+    if (m_networkAccessManager)
+        m_networkAccessManager->deleteLater();
 }
 
 void QSampleCache::loadingRelease()
